Application Code/Myproject/requirements.txt
absl-py==0.11.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.1
astunparse==1.6.3
async-generator==1.10
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
channels==3.0.2
chardet==3.0.4
click==7.1.2
cmdstanpy==0.9.5
colorama==0.4.4
configparser==5.0.1
constantly==15.1.0
convertdate==2.3.0
crayons==0.4.0
cryptography==3.2.1
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.21
daphne==3.0.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.3
django-cors-headers==3.5.0
django-rest-knox==4.1.0
django-suit==0.2.28
djangorestframework==3.12.2
entrypoints==0.3
ephem==3.7.7.1
fake-useragent==0.1.11
filelock==3.0.12
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
holidays==0.10.3
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
ipykernel==5.3.4
ipython
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==21.4.0
kiwisolver==1.2.0
korean-lunar-calendar==0.2.1
lerc==0.1.0
lightgbm==3.1.0
LunarCalendar==0.0.9
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
nltk==3.5
notebook==6.1.4
ntlm-auth==1.5.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.3
parse==1.18.0
parso==0.7.1
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyee==7.0.4
Pygments==2.7.1
PyHamcrest==2.0.2
PyMeeus==0.3.7
pyOpenSSL==20.0.0
pyparsing==2.4.7
pypi==2.1
pyppeteer==0.2.2
pyquery==1.4.3
pyrsistent==0.17.3
pyshp==2.1.2
PySocks==1.7.1
pystan==2.19.1.1
python-dateutil==2.8.1
pytz==2020.1
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
realtime-stock==1.0.0
regex==2020.11.13
requests==2.24.0
requests-html==0.10.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
setuptools-git==1.2
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
sqlparse==0.4.1
stock-info==1.0.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tflearn==0.5.0
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.52.0
traitlets
tweepy==3.9.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.11
virtualenv==20.2.0
w3lib==1.22.0
wcwidth==0.2.5
webdriver-manager==3.2.2
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.2.1
gunicorn==20.0.4
python-dotenv
yahoo-fin==0.8.6
zope.interface==5.2.0

